# api/openapi.yaml
openapi: "3.1.0"
info:
  title: SoroTrail API
  description: |
    A contract event indexer for the Stellar/Soroban network. SoroTrail polls
    a Stellar RPC endpoint, stores contract events durably in Postgres, and
    serves them back through a queryable HTTP API long after the RPC has
    forgotten them.
  version: "0.1.0"
servers:
  - url: http://localhost:8080
    description: Local development
# Authentication is instance-dependent, so the default requirement lists the
# tenant credential alongside the empty requirement: a MULTI_TENANT=true
# deployment rejects an anonymous caller with 401, and a single-tenant one
# serves it. Operations that are gated regardless override this with their
# own security block.
security:
  - TenantAuth: []
  - TenantAuthHeader: []
  - {}
paths:
...
  /apikeys:
    get:
      operationId: listAPIKeys
      summary: List API keys
      description: >-
        Returns the metadata for every API key on this instance. The
        plaintext key is never returned — it is visible only in the
        response to POST /apikeys, at creation time.

        This endpoint is authenticated regardless of API_KEY_AUTH_ENABLED:
        an unauthenticated key-management surface would let any caller mint
        a key and bypass authentication entirely. The first key is issued
        out of band with `sorotrail apikey`.
      tags: [Auth]
      security:
        - TenantAuth: []
        - TenantAuthHeader: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKey"
              example:
                - id: 1
                  name: "ci-reader"
                  prefix: "a1b2c3d4e5f6g7h8"
                  created_at: "2026-09-01T12:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createAPIKey
      summary: Create an API key
      description: >-
        Mints a new API key and returns it. The `key` field in the response
        is the only time the plaintext key is ever available — only a
        bcrypt hash of its secret half is persisted, so a lost key cannot
        be recovered and must be revoked and reissued.

        The request body is optional; omitting it creates an unnamed key.
      tags: [Auth]
      security:
        - TenantAuth: []
        - TenantAuthHeader: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
                  description: >-
                    A human-readable label for the key, to make it
                    identifiable in the list response. At most 100
                    characters.
              example:
                name: "ci-reader"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIKey"
                  - type: object
                    required: [key]
                    properties:
                      key:
                        type: string
                        description: >-
                          The plaintext key, in the form
                          `st_<prefix>_<secret>`. Returned once and never
                          again; store it before discarding the response.
              example:
                id: 1
                name: "ci-reader"
                prefix: "a1b2c3d4e5f6g7h8"
                created_at: "2026-09-01T12:00:00Z"
                key: "st_a1b2c3d4e5f6g7h8_Zm9vYmFyYmF6cXV4..."
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /apikeys/{id}:
    delete:
      operationId: revokeAPIKey
      summary: Revoke an API key
      description: >-
        Marks the key revoked. The row is kept so the key's history stays
        auditable, but the key stops authenticating immediately. Revoking
        an already-revoked key is not an error.
      tags: [Auth]
      security:
        - TenantAuth: []
        - TenantAuthHeader: []
      parameters:
        - name: id
          in: path
          required: true
          description: >-
            The key's database ID, as returned by POST /apikeys — not its
            public prefix and not the plaintext key. Must be a positive
            integer.
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 1
      responses:
        "204":
          description: Revoked
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  schemas:
    APIKey:
      type: object
      description: >-
        API key metadata. The secret half of the key is never included —
        only a bcrypt hash of it is stored, and the plaintext is returned
        exactly once, by POST /apikeys.
      properties:
        id:
          type: integer
          format: int64
          description: Database ID, used to revoke the key.
        name:
          type: string
          description: Human-readable label, empty for an unnamed key.
        prefix:
          type: string
          description: >-
            The non-secret 16-character lookup handle stored in the clear.
            It identifies the key without revealing it.
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          description: >-
            When the key was revoked. Absent while the key is active.
      required: [id, name, prefix, created_at]
    ErrorResponse:
      type: object
      description: Standard error envelope
      properties:
        error:
          type: string
          description: Human-readable error message
      required: [error]
...
  securitySchemes:
    TenantAuth:
      type: http
      scheme: bearer
      description: |
        The tenant API key issued by POST /admin/tenants/{id}/keys, presented
        as `Authorization: Bearer st_<prefix>_<secret>`. The same key is also
        accepted in an `X-API-Key` header (see TenantAuthHeader); Bearer wins
        when both are present.

        Only enforced when the instance runs with MULTI_TENANT=true. On a
        single-tenant instance every request is served as an implicit
        wildcard principal and no credential is required, which is why most
        operations list this scheme as optional rather than required.
    TenantAuthHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        The same tenant API key as TenantAuth, presented in a header instead
        of a Bearer token.
    ManagementKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        The operator's instance-wide management key, configured through the
        API_KEY environment variable. It gates the endpoints that mutate
        ingestion or read the dead-letter queue, and it is always required on
        those endpoints — independently of MULTI_TENANT, and independently of
        the tenant credential above.

        When API_KEY is unset the server fails closed: those endpoints answer
        503 rather than running unauthenticated.
  responses:
...
    BadRequest:
      description: >-
        The request could not be parsed or violates a documented constraint —
        a malformed identifier, an out-of-range limit, an unparseable cursor,
        or a JSON body the endpoint cannot decode. The message names the
        offending parameter.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: |
        No credential was presented, or the one presented is malformed,
        unknown or revoked.

        Which credential is missing depends on the endpoint's security
        requirement: endpoints requiring ManagementKey want the operator's
        `X-API-Key`, while everything else wants the tenant key and only
        rejects when the instance runs with MULTI_TENANT=true.
      headers:
        WWW-Authenticate:
          description: >-
            The authentication challenge, per RFC 7235 §4.1 — for example
            `Bearer realm="sorotrail"`. Present on the tenant-authentication
            failures; the ManagementKey path answers without a challenge
            because it accepts no Bearer token.
          schema:
            type: string
            examples: ['Bearer realm="sorotrail"']
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: >-
        The addressed resource does not exist, or is not visible to the
        caller. Event lookups answer 404 rather than 403 for an event the
        caller may not read, so the endpoint cannot be used as an existence
        oracle. The tenant and admin endpoints also answer 404 when the
        instance runs single-tenant and those routes are therefore inert.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequests:
      description: |
        The caller exceeded its request quota, or asked to watch more
        contracts than its quota allows.

        Rate limiting is per client — keyed on the authenticated tenant where
        one exists and on the client IP otherwise — and is only active when
        the operator configured RATE_LIMIT_RPS or a per-tenant override. The
        probe and scrape endpoints (/health, /livez, /readyz, /metrics) are
        never rate limited, so throttling cannot blind an operator to the
        throttling itself.
      headers:
        Retry-After:
          description: >-
            Whole seconds to wait before retrying, per RFC 7231 §7.1.3.
            Always at least 1, and rounded up from the limiter's own
            wait-to-next-token estimate.
          schema:
            type: integer
            minimum: 1
            examples: [1]
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: >-
        The request failed for a reason the caller cannot act on — a store or
        RPC failure, or a panic caught by the recovery middleware. The
        message is deliberately generic; the specifics are logged against the
        request's X-Request-ID rather than returned.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...